refactoring form and update test
