+++ src/Entity/Category.php
@@ -42,11 +42,4 @@ class Category
 
         return $this;
     }
-
-    public function __toString(){
-        // to show the name of the Category in the select
-        return $this->name;
-        // to show the id of the Category in the select
-        // return $this->id;
-    }
 }
