+++ src/Form/LinkType.php
@@ -2,7 +2,9 @@
 
 namespace App\Form;
 
+use App\Entity\Category;
 use App\Entity\Link;
+use Symfony\Bridge\Doctrine\Form\Type\EntityType;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
@@ -14,7 +16,10 @@ class LinkType extends AbstractType
         $builder
             ->add('description')
             ->add('link')
-            ->add('category')
+            ->add('category', EntityType::class,[
+                'class' => Category::class,
+                'choice_label' => 'name'
+            ])
         ;
     }
